Fixed CGFloat type conversion in Extensions.swift to remove warning

+++ Angelica Fighti/Extensions.swift
@@ -107,7 +107,7 @@ extension SKLabelNode{
 /*RANDOM FUNCTIONS */
 
 func random() -> CGFloat {
-    return CGFloat(Float(arc4random()) / 0xFFFFFFFF)
+    return CGFloat(Double(arc4random()) / 0xFFFFFFFF)
 }
 
 func random( min: CGFloat, max: CGFloat) -> CGFloat {
